Restructures app and premium component registration
